[wip] Convert ReactNative to use CKTextKit
Provides asynchronous layout and rendering with a much safer, and
capable text engine.

The core differences here are in the fact that CKTextComponentLayer
provides full asynchronous rendering of text content, and that all text
layout artifacts are accessed via a central, thread safe LRU cache.
This keeps the memory profile stable no matter how many text elements
are rendered in an enormous table view.  It also caches raster layer
contents so if you’re scrolling back over some text that you’ve already
seen, it can usually just use the cache instead of re-rendering.

This is the same rendering engine that we wrote for ComponentKit, and
closely mirrors the internal text view we use at FB.